eth, polygon/bor: fix fetch span logic for devnets (#9312)
This PR fixes 2 things which are more commonly visible in multi client
devnet setups for polygon.

Context: The span logic is a bit different when it comes to first 2
spans. Bor/Erigon makes the first commit for a span during the start of
sprint (i.e. if sprint length is 16, it will call `commitSpan` at block
16 in bor consensus for the first time). Span 0 is hard coded in genesis
contracts so it needs to commit span with `id=1` on that block (see
equivalent code in bor
[here](https://github.com/maticnetwork/bor/blob/v1.2.3/consensus/bor/bor.go#L1150-L1152)).
At that time, it needs to have 1st span available. Bor fetches it on the
go while erigon processes it in a separate stage and stores in a
snapshot. Hence, we'd need to fetch 1st span as an exception in erigon
while we're still in span 0 but also need to make sure that it doesn't
block processing of any previous blocks.

Based on https://github.com/ledgerwatch/erigon/pull/9149, the span ID
used to fetch and commit span was wrong and the span 1 needs to be
loaded explicitly in 1st sprint.