Remove dead transaction time code (#9563)
The `core/types/` transaction logic contains bits that no longer offer
any function. This commit removes the 'time', which the documentation
states is for spam protection on when a tx is first seen. Since this is
all handled inside the txpool now, the 'time' is never actually
populated and is simply unnecessary code. Similarly, the code for
sorting transactions by time is likewise never invoked.