New overlays feature for doing ad-hoc simulations of existing contracts with modified bytecode (#9438)
In this PR I'm introducing so-called `overlays` for doing ad-hoc state
simulations of contracts which replay the state with a modified bytecode
for any given block range.

The new feature adds two new RPC methods: `overlay_callConstructor` and
`overlay_getLogs`

I've also added a few bench tests and a new postman collection which
checks both new methods for the WETH contract in original and modified
version.

## Motivation
Overlays allow you to add your custom logic to already deployed
contracts and simulate on top of them. With overlays you can create new
view functions, modify existing ones, change field visibility, emit new
events and query the historical data of any contract with your modified
source code.

Similar commercial products exist in the space and I believe making this
feature opensource in erigon will make it easier for everyone to tinker
with it and build new cool things on top of it :sparkles:

## Usage
1) compile your contract to obtain the new deployment bytecode, append
the constructor args at the end, if any
2) call `overlay_callConstructor` with the result from 1) to receive the
runtime bytecode
3) call `overlay_getLogs` to simulate new logs, passing 2) as state
overrides
([example](https://github.com/crebsy/erigon/blob/devel/cmd/rpcdaemon/postman/Overlay_Testing.json))
4) call `eth_call` to simulate new view function calls passing 2) as
state overrides

## Ad-hoc query performance & historical back-filling
The current implementation for ad-hoc range queries of `overlay_getLogs`
is optimized so that irrelevant blocks are skipped entirely from the
simulation context. This can lead to very fast results of a few minutes
or even seconds for simulating the entire historical logs of a contract
which makes any infra for back-filling redundant. However, such
shortcuts are only effectively working for contracts which are not
touching too many blocks of the chain's history. So depending on the
contract you're trying to patch, you still might need to create a custom
back-filling script which simulates the new logs with e.g. 1000 or 10000
block ranges.

As a reference for a "busy" contract the simulation rate for the
[Uniswap V2 router
](https://etherscan.io/address/0x7a250d5630b4cf539739df2c5dacb4c659f2488d#code)
was ~ 280 blocks/s on a host with 24 cores. So while small block ranges
returned relatively quickly, the simulation of the entire historical
logs for this contract took a few hours on a single node.

I'm not collecting any perf metrics yet for this but this could be
easily added in the future for tracking better implementations and
insights about contract <-> block distribution.

## Configuration
1) Enable the `ots,overlay` api namespaces by adding it to the
`--http.api` flag

2) The implementation of `overlay_getLogs` is parallelized so it's good
practice to increase the `--db.read.concurrency` to a higher value, e.g.
256 so other erigon goroutines are not getting blocked.


## Supported networks
Tested with ethereum and polygon nodes