Caplin: Refactored state antiquary (#9876)
This PR separates the data collection layer from the `Antiquary` object
to the `beaconStatesCollector`, which hopefully makes the code cleaner.
It is still messy, but now we have a proper separation between the logic
and the data collection layer.

The reason why we have internal buffers here is that MDBX locks the
entire database when writing to it, so we need to minimize the time
spent in the write lock. So instead of writing the historical states on
write transactions and keeping the lock for hours, we accumulate them in
memory and write them in a single write transaction.