p2p/sentry: StatusDataProvider ReadCurrentHeader error (#9890)
P2P fails on restart because rawdb.ReadCurrentHeader returns a nil
header. It looks like ReadHeadHeaderHash fails to find the current
header hash. However the correct hash is returned by ReadHeadBlockHash.

Let's use ReadHeadBlockHash, because the status needs to report a header for which we have a full block body.