polygon/heimdall: refactor into txReadStore and txStore (#10130)
Will be using `NewRwTxStore` in
https://github.com/ledgerwatch/erigon/pull/10124

Idea is simple - split the current store in 2:
- `roTxStore`
- `rwTxStore`

These 2 stores wrap around a tx with snapshots reader. 
The `rwTxStore` wraps around the `roTxStore` for code re-use.
This split removes chances of unexpected `tx is read only` errors when
calling the store update methods.

Also, opens up a possibility to introduce a `dbRwStore` and `dbRoStore`
wrappers around `roTxStore` and `rwTxStore` if we decide to go down this
path. The `dbRwStore` in particular can have ownership over the lifetime
of a `rwTxStore` - make sure only 1 is open at a time for example by
using a channel/mutex/semaphore or some other idea and call `Rollback`
and/or `Commit`.

We may need to look into transitioning these to `TemporalTx` at some
point - for now let's just get this to work first.