Make E35 block production produce good blocks (#10133)
This PR makes Erigon produce empty blocks when running with E35.

It can include transactions, withdrawals and blobs. I do not know if it
fits well with Polygon.

Example of block produced by E35:
https://sepolia.etherscan.io/block/5809362.

Approach is simple. have `ExecV3` compute the commitment, that was the
only difference. it requires to simulate block execution + state root
computation with a skip on state root check.