polygon/heimdallsim: refactor HeimdallSimulator (#10433)
This is a simplification refactoring in preparation of using this code
for a `heimdall.Service` integration test.

* remove extra features: downloading data using a torrent client,
treating cfg.Preverified files. Now it only works with the .seg files
available in the passed directory;
* fix conventions: naming, ctx usage;
* avoid a grpc dependency;
* make sure HeimdallSimulator implements HeimdallClient interface;
* add a `.Next()` method to control iterations (note: in the future
we'll need a separate Next method for each entity)
* move the package to devnet polygon services. Reasons for moving:
1) HeimdallSimulator is going to be used for integration tests and
devnet is our current place for those tests;
2) `devnet/service/polygon` already contains a Heimdall simulator class
that also implements HeimdallClient interface and generates data
dynamically;
3) it was confusing to me that simulator was a direct subpackage of
heimdall, because simulator depends on heimdall. Usually a direct
subpackage is a dependency of the parent. IMO a package shouldn't depend
on an ancestor package (it can depend on a separate branch of the dir
tree, but not the same branch).