BodyDownload: differentiate between empty vs nil withdrawals/requests (#10436)
With withdrawals we differentiate between missing (`nil`) and
present-but-empty cases. Pre-Shapella blocks should have no withdrawals
(i.e. `nil`), while post-Shappela blocks must have withdrawals, even if
empty. The same goes for
[EIP-7685](https://eips.ethereum.org/EIPS/eip-7685) requests pre- and
post-Pectra.

`BodyHashes` is used to map downloaded block bodies to corresponding
headers. Prior to this change, it was possible to have a mismatch
between block & header in terms of `nil` vs empty withdrawals or
requests because both `nil` & empty were coded as an `EmptyRootHash`.
With this change only empty withdrawals/requests are coded as an
`EmptyRootHash`, while `nil` ones are coded as zeros, thus ensuring that
blocks & headers are consistent in terms of `nil` vs empty. (This
invariant is also checked in `(b *Block) HashCheck()`.)