stagedsync: polygon sync stage to append canonical tx nums (#10574)
While testing the new polygon sync stage I run into the below error:
```
[EROR] [05-29|21:26:54.099] Staged Sync                              err="[4/6 Execution] ExecV3: TxNums index is at block 6399999 and behind commitment 6426107: behind commitment"
```

This is because we weren't appending the canonical tx numbers to the
TxNums index.

This PR fixes this by extending `fixCanonicalChain` to return `newNodes`
and `badNodes` on the path to the connecting point from the new fork
choice update. `newNodes` can be used to append the right tx nums to the
tx nums index while the `badNodes` can be used to unwind (this will be
properly handled in a subsequent PR).