Raise the limit of debug_accountRange to 8192 (if storage not requested) (#10753)
`debug_accountRange` is one of the cool rpc methods erigon provides.
However, the hardcoded limit 256 makes it supper inefficient. In my
personal benchmark, raising the limit to 8192 could boost the indexing
process 3-4x faster and thus this PR raises the default limit.

I could understand the security consideration of the DDoS attack on this
parameter. But it's an RPC method under the `debug` namespace, which is
not common for external usage, and both limits impose similar database
pressure. Therefore, it should be mostly safe to raise the limit to a
moderate value.