bplus index inefficient warmup (#10889)
2 bugs inside and outside `bpsTree.traverse`:

- depth of the tree was always 0 which created flat PARTIALLY ordered
list of keys. Also 2 useless keys (first and second) were permanently
added into cache.
- bug in iteration caused less node to be inserted in cache, so we did
more random reads to files.

If set `cacheEvery` to 4 or 8 it would take each `x` element into cache.
if `=2` then ~0.8% of keys are in cache. `cacheEvery=8` keeps ~3.3% of
keys in memory and slightly decreases amount of cow and splits in db,
and i think this trade of doesnt worth it.
Feels like having several layers of lists in cache does not increase
speed but increases logic complexity

This change does not require creation of new indices, but seriously
affects structure of in-mem cache. Causes less random reads as well.

---------

Co-authored-by: alex.sharov <AskAlexSharov@gmail.com>