reset sigint default handler after libp2p.New call (#11694)
After investigation I see that SIGINT default handler is modified AFTER
the line `host, err := libp2p.New(opts...)`

I looked into the library but couldn't find where the handlers are
modified. From experimentations it's clear that this is the line after
which the SIGINT behaviour changes (does nothing).

I'm resetting the default behavior after the call. 

I tested it simply running the caplin for a few seconds and shutting it
down with ctl+c which didn't work before the change.

---------

Co-authored-by: shota.silagadze <shota.silagadze@taal.com>