downloader: added set log prefix grpc method (#11731)
To print the current download progress, we currently call the GRPC
method RecalcStats for every log entry. However, given the frequent
updates to the downloader in recent days, we may need to add more
information to the logs. This would require updating the Go interfaces,
which can be painful.

To streamline this process, I've added a log prefix to move the
responsibility of logging the download progress directly to the
downloader. This change also allows us to print a proper prefix
(currently "OtterSync"). As a result, we may soon be able to remove
StatsReply, since Erigon is primarily concerned with whether the
download is complete or not.