Rework goreleaser configuration, release workflow and Dockerfile for release images. (#11766)
It is a work on https://github.com/erigontech/erigon/issues/10251 .

Main changes:
- goreleaser reconfigured (more binaries added, most of the build flags
passed to goreleaser)
- released erigon binary now has "Build info" dynamic values such as a
branch, tag, commit id.
- amd64 binaries will be build using GOAMD64=v2 support for better
support of modern CPUs.
- as agreed -- now it will generate release without prefix "v", i.e.
"2.60.7", but not "v2.60.7".
- docker image will have following binaries: erigon, integration, diag,
sentry, txpool, downloader and rpcdaemon. (Exactly the same binaries as
published in artifacts in the release notes).
- multi-platform docker tag created with support of two OS/Arch --
linux/amd64/v2 and linux/arm64 which should cover almost all use cases
when docker container used.
- docker images will have labels which refer to full commit id and short
commit id, build timestamp, and others. It may help to identify which
commit is a part of the image.
- this workflow could be executed in a "dry_run" mode, i.e. to build
artifacts and keep it for a 1d in a workflow artifact store (in this
case we can download artifacts and test before actual release).

Windows artifacts will be not released for now (require fix).