Optimise aggregate and proof signature verification (#11816)
We are trying to optimise `AggregateAndProofService`. After profiling
the service, I see that most of the CPU time is spent on signature
verifications. From the graph, overall, the function took 6.6% (74
seconds) of all the time (not just execution time. Percentage would be
much higher if we took just cpu time) see the screenshot:

<img width="1470" alt="Screenshot 2024-09-01 at 10 28 38"
src="https://github.com/user-attachments/assets/929ce103-2bf3-43d9-a0fa-ca504e4b58bb">


Now we are trying to aggregate all the signatures and verify them
altogether with `bls.VerifyMultipleSignatures` function in an async way
and run the final functions if verifications succeed. I basically
removed all the code where we verified those three signatures and
instead gathered them for verifying later. After profiling that I see
the following output:

<img width="1468" alt="Screenshot 2024-09-01 at 10 44 31"
src="https://github.com/user-attachments/assets/abb842a3-0b4f-4640-8a88-791a2d0af62b">

Now most of the time, as I see, is spent on public key aggregation when
we are verifying validator aggregated signatures. But I guess there is
no way to optimise that one. As we are spending most of the time on
`NewPublicKeyFromBytes` maybe we could cache constructed keys but I
think bls already does that. So, that's as good as it gets.

---------

Co-authored-by: shota.silagadze <shota.silagadze@taal.com>