polygon/sync: block downloader to not re-insert blocks behind start (#11929)
Run into an issue since we started pruning total difficulty.
```
EROR[09-09|10:58:03.057] [2/6 PolygonSync] stopping node          err="parent's total difficulty not found with hash 9334099de5d77c0d56afefde9985d44f8b4416db99dfe926908d5501fa8dbd9e and height 11736178: <nil>
```

It happened for checkpoint
[9703](https://heimdall-api-amoy.polygon.technology/checkpoints/9703).
Our start block was in the middle of the checkpoint range which meant we
have to fetch all the 8k blocks in this checkpoint to verify the
checkpoint root hash when receiving blocks from the peer.

The current logic will attempt to insert all these 8k blocks and it will
fail with a missing parent td error because we only keep the last 1000
parent td records.

This PR fixes this by enhancing the block downloader to not re-insert
blocks behind the `start` block. This solves the parent td error and
also is saving some unnecessary inserts on the first waypoint processing
on startup.