Make findEpochRoot work with missing blocks at the end of an epoch
This is for example used in attestations rewards calculation, which
was failing with the error message "no block found for this epoch".

In reality only the last block was not found in the epoch, and that
already triggered the error.

With this change, we find the last non-missed block in this case.