Fix attestations reward calculation inaccuracy
This typo (epoch vs. prevEpoch) caused little calculation differences
compared to reality (as shown by balance changes on eth2).

Unfortunate, that we didn't have some kind of type system check catch
this, as debugging took 2 days, and the name of the parameter showed
that the function wants the previous epoch and not the current one.

Maybe previous epoch and epoch shouldn't be the same type...