polygon/sync: avoid dropping milestone events (#12269)
Fixes an issue with Astrid at tip where the root of our canonical chain
builder got too far away from the tip (> 1024 blocks) which is not to be
expected. It should never happen at tip. Milestones are usually ~30
blocks on Mainnet. I've seen worst cases around ~200 blocks on Amoy. So
if our processing is accurate our root should never be more than 1024
aways from the tip. The explanation for this is that every time we
process a milestone we move our root forward, so it should always be 1
milestone worth of blocks behind tip.

The underlying root cause of this caused 2 problems:
1. an error that crashed the node 
```
[EROR] [10-09|16:54:25.194] [2/6 PolygonSync] stopping node          err="invalid fetch blocks amount: amount=1030"
```

<br/>

2. degrading performance of the node for some time before the error

![Screenshot 2024-10-10 at 16 44
37](https://github.com/user-attachments/assets/f51afc7e-ba32-4619-a56f-b2cb3b59a1a4)


The underlying issue causing the above is that we've dropped milestone
events from our `EventsChannel` (it has a capacity of 1000 events and
drops oldest events when pushing an event at capacity). We can confirm
this via the logs. Notice the gap between the end block of the first and
the start block of the second - these are 2 consecutive events that we
have processed so there should be no gaps and the time between them
should not be ~4mins (should be ~1 min).
```
[DBUG] [10-09|16:14:43.079] [sync] applying new milestone event      milestoneStartBlockNum=62831140 milestoneEndBlockNum=62831154 milestoneRootHash=0xebccef3890f3ba64d1bec949c5ce298acc88bca5e333520546b7dabc473b9043
...
[DBUG] [10-09|16:19:01.605] [sync] applying new milestone event      milestoneStartBlockNum=62831259 milestoneEndBlockNum=62831274 milestoneRootHash=0xac2f53dc691a63e2672ee85572f29670c6217df8cfd68796c680189b08156652
```

<br/>

Additionally, the heimdall process logs confirm we have scrapped the
milestones. Milestone 494026 corresponds to the last milestone event we
have processed above (ending at block 62831274) before the error.
Milestone 494027 is next and we've scrapped it ~36 seconds later.
However we've never processed it in Astrid leading to our root falling
behind too much.
```
INFO [2024-10-09|16:17:41.841] Served RPC HTTP response                     module=rest-server method=GET url=/milestone/494026 status=200 duration=0 remoteAddr=127.0.0.1:46790
...
INFO [2024-10-09|16:18:17.452] Served RPC HTTP response                     module=rest-server method=GET url=/milestone/494027 status=200 duration=1 remoteAddr=127.0.0.1:49038
```

<br/>

From all of this is clear that we're dropping milestone events in our
EventsChannel. I spent time investigating why. When we connect to many
peers on mainnet (think >50 peers) we start receiving a lot of NewBlock
and NewBlockHashes events from devp2p. If the amount of devp2p events
coming in is higher than the rate at which we consume/process the events
then we will start dropping stuff from the events channel.

This is fine in normal circumstances however Erigon (non-astrid) peers
seem to over-propagate NewBlockHashes. This is due to the logic in the
sentry multi client - when a NewBlock event is received, Erigon responds
with NewBlockHashes announcements to all of its peers - without keeping
track if that peer already knows about this hash (either because Erigon
has already announced that block hash or Erigon has received that block
hash announcement from the peer). The end result of this is that the
more Erigon peers we connect to the more we will get flooded. This
causes a lot of unnecessary processing for us in Astrid and also in us
dropping milestone events. Additionally, it is also not following the
devp2p specs
https://github.com/ethereum/devp2p/blob/master/caps/eth.md#newblockhashes-0x01.
*NOTE block propagation via sentry multi client is disabled in Astrid,
so it does not suffer from this problem. Block propagation for Astrid
will be implemented in a follow up PR as part of the driver and it would
not suffer from this problem.*

The above is confirmed by the following logs: 

[period-of-event-spam-log.txt](https://github.com/user-attachments/files/17331131/period-of-event-spam-log.txt).

In the logs, notice how many duplicate events for the same block we have
received from peer
`b612fe4679bc08112367904fbb00b26ba9ff1557f7c9c597fc0387a417637a36e6a5f897d2c6bf76a3843e16da286ae949ed11f456228460fff26189bc13fceb`
and peer
`096ba1be0fbc2996d966a1fce41e309d887ccf322e9a9ffc252b722dc0c65ed0ff4998d2b99671cfdba1311e607860d414e49aa05176dc9bca49d67596498480`.
These are both Erigon (non-Astrid) peers. Confirmed via the diagnostics
tool.

The solution to all of this in this PR is:
1. Introduce a new `CompositeEventsChannel` which allows us to split our
`EventsChannel` into 2 - one for heimdall milestone events so they do
not get dropped by the devp2p messages and one for the devp2p messages.
The `CompositeEventsChannel` unifies them so we have to deal with only 1
channel.
2. Introduce a `blockEventsSpamGuard` to protect us from filling our
`EventsChannel` with loads and loads of duplicate devp2p messages from
the same peer